progress on behavioral template
